requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-telegram-bot>=20.0
openai>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
python-dotenv>=1.0.0
